improve code organization
