+++ rogueliketcodtutorial2021/__main__.py
@@ -1,33 +1,6 @@
 import sys
 
-import tcod
-
-from rogueliketcodtutorial2021.rogueliketcodtutorial2021 import fib
-
-
-def main() -> None:
-    screen_width = 80
-    screen_height = 50
-
-    tileset = tcod.tileset.load_tilesheet("dejavu10x10_gs_tc.png", 32, 8, tcod.tileset.CHARMAP_TCOD)
-
-    with tcod.context.new_terminal(
-        screen_width,
-        screen_height,
-        tileset=tileset,
-        title="Yet Another Roguelike Tutorial",
-        vsync=True,
-    ) as context:
-        root_console = tcod.Console(screen_width, screen_height, order="F")
-        while True:
-            root_console.print(x=1, y=1, string="@")
-
-            context.present(root_console)
-
-            for event in tcod.event.wait():
-                if event.type == "QUIT":
-                    raise SystemExit()
-
+from rogueliketcodtutorial2021.rogueliketcodtutorial2021 import fib, main
 
 if __name__ == "__main__":
     if len(sys.argv) >= 2:
