+++ rogueliketcodtutorial2021/rogueliketcodtutorial2021.py
@@ -1,5 +1,35 @@
+import tcod
+
+
 def fib(n: int) -> int:
     if n < 2:
         return n
     else:
         return fib(n - 1) + fib(n - 2)
+
+
+def main() -> None:
+    screen_width = 80
+    screen_height = 50
+
+    player_x = int(screen_width / 2)
+    player_y = int(screen_height / 2)
+
+    tileset = tcod.tileset.load_tilesheet("dejavu10x10_gs_tc.png", 32, 8, tcod.tileset.CHARMAP_TCOD)
+
+    with tcod.context.new_terminal(
+        screen_width,
+        screen_height,
+        tileset=tileset,
+        title="Yet Another Roguelike Tutorial",
+        vsync=True,
+    ) as context:
+        root_console = tcod.Console(screen_width, screen_height, order="F")
+        while True:
+            root_console.print(x=player_x, y=player_y, string="@")
+
+            context.present(root_console)
+
+            for event in tcod.event.wait():
+                if event.type == "QUIT":
+                    raise SystemExit()
